test: move format_{string,luav} to a separate module
